+++ src/index.js
@@ -2,3 +2,4 @@
 // book. Import a subpath (`book-of-elementals/accordion`) to register just one.
 export * from './core.js';
 export * from './elementals/accordion/index.js';
+export * from './elementals/disclosure/index.js';